backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
google-auth
google-auth-oauthlib
google-auth-httplib2
httpx
python-dotenv
pydantic
pydantic-settings
email-validator
fastapi-mail
sqlalchemy
Pillow
httpx

# ─── AI Moderation — Couche 1 : détection toxicité texte (BERT) ───────────────
detoxify>=0.5.2

# ─── AI Moderation — Couche 2 : classification image NSFW (CNN) ───────────────
nudenet>=3.4.2

# ─── AI Moderation — Couches 2+3 : NLP + Vision IA ───────────────────────────
# XLM-RoBERTa zero-shot (classification texte multilingue FR/AR/EN)
# CLIP ViT-B/32 (analyse sémantique image — détecte accidents, hors-sujet, etc.)
transformers>=4.40.0
torch>=2.1.0
torchvision>=0.16.0
sentencepiece>=0.1.99      # Tokenizer XLM-RoBERTa
protobuf>=4.0.0            # Requis par SentencePiece pour XLM-RoBERTa
tiktoken>=0.5.0            # Fallback tokenizer (BPE)

# ─── AI Moderation — Couche 5 : CNN custom (Text CNN + ResNet18) ───────────
pandas>=1.5.0
scikit-learn>=1.3.0
